Fix that some more

+++ internal/caching/cache_roomservernids.go
@@ -25,9 +25,9 @@ type RoomServerNIDsCache interface {
 }
 
 func (c Caches) GetRoomServerRoomID(roomNID types.RoomNID) (string, bool) {
-	return c.RoomServerRoomIDs.Get(roomNID)
+	return c.RoomServerRoomIDs.Get(int64(roomNID))
 }
 
 func (c Caches) StoreRoomServerRoomID(roomNID types.RoomNID, roomID string) {
-	c.RoomServerRoomIDs.Set(roomNID, roomID)
+	c.RoomServerRoomIDs.Set(int64(roomNID), roomID)
 }
